refactor(metrics): extract planner_metrics as a shared workspace library
Relocate the subscore/geometry/config code out of the diffusion_planner package
into a new bottom-layer workspace member `planner_metrics`, so the shared metrics
library no longer lives inside one of its consumers. Both diffusion_planner and
rlvr now depend on it.

- New member planner_metrics/ (deps: numpy, torch only — no rlvr/diffusion_planner
  dependency): the moved config/geometry/subscores/aggregate modules + a new
  collision_geometry.py holding the two pure OBB primitives
  (batch_signed_distance_rect / center_rect_to_points), moved verbatim from
  model/guidance/collision.py.
- diffusion_planner.model.guidance.collision re-exports those two primitives from
  planner_metrics (shim) so every existing call site keeps working.
- rlvr.reward and the reward tests import from planner_metrics instead of the old
  diffusion_planner.metrics (removed).
- Workspace wiring: planner_metrics added as a member + source; diffusion_planner
  and rlvr declare it as a workspace dependency.

Pure relocation, no behavior change: the two collision primitives are
byte-identical (ast-verified), golden tests prove compute_reward_batch output is
unchanged, and parity + full reward + guidance suites are green (158 passed).
Resolves the layering issue where a cross-cutting metrics library lived under the
diffusion_planner package.
