--- diffusion_planner/diffusion_planner/metrics/__init__.py
@@ -1,9 +1,0 @@
-"""Neutral metrics library: raw EPDMS-style subscores + geometry + config,
-importable without depending on ``rlvr``."""
-
-from __future__ import annotations
-
-from diffusion_planner.metrics.aggregate import compute_subscores_batch  # noqa: F401
-from diffusion_planner.metrics.config import RewardConfig  # noqa: F401  (re-export)
-from diffusion_planner.metrics.geometry import *  # noqa: F401,F403
-from diffusion_planner.metrics.subscores import *  # noqa: F401,F403
