+++ diffusion_planner/pyproject.toml
@@ -14,6 +14,7 @@ dependencies = [
     "opencv-python",
     "pandas>=2,<3",
     "peft",
+    "planner-metrics",
     "protobuf",
     "pydantic",
     "pytorch-lightning",
@@ -27,3 +28,6 @@ dependencies = [
 [build-system]
 requires = ["setuptools>=61"]
 build-backend = "setuptools.build_meta"
+
+[tool.uv.sources]
+planner-metrics = { workspace = true }
