+++ planner_metrics/__init__.py
@@ -1,0 +1,9 @@
+"""Neutral metrics library: raw EPDMS-style subscores + geometry + config,
+importable without depending on ``rlvr``."""
+
+from __future__ import annotations
+
+from planner_metrics.aggregate import compute_subscores_batch  # noqa: F401
+from planner_metrics.config import RewardConfig  # noqa: F401  (re-export)
+from planner_metrics.geometry import *  # noqa: F401,F403
+from planner_metrics.subscores import *  # noqa: F401,F403
