+++ rlvr/pyproject.toml
@@ -11,6 +11,7 @@ dependencies = [
     "matplotlib",
     "numpy>=1.26,<2",
     "pandas>=2,<3",
+    "planner-metrics",
     "preference-optimization",
     "safetensors",
     "scenario-generation",
@@ -37,6 +38,7 @@ rlvr = ["configs/*.json", "*.npy"]
 
 [tool.uv.sources]
 diffusion-planner = { workspace = true }
+planner-metrics = { workspace = true }
 preference-optimization = { workspace = true }
 exploration-policy = { workspace = true }
 guidance-gui = { workspace = true }
